Fixed test and moved utilities
